Improve back-end API using MVC framework
